+++ src/Classes/MathUtil.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contao Utilities for Contao Open Source CMS
+ * Copyright (c) 2019-2023 Web ex Machina
+ *
+ * @category ContaoBundle
+ * @package  Web-Ex-Machina/contao-utils
+ * @author   Web ex Machina <contact@webexmachina.fr>
+ * @link     https://github.com/Web-Ex-Machina/contao-utils/
+ */
+
+namespace WEM\UtilsBundle\Classes;
+
+class MathUtil
+{
+    /**
+     * Calculates the great-circle distance between two points, with the Vincenty formula.
+     *
+     * @param float $latitudeFrom  Latitude of start point in [deg decimal]
+     * @param float $longitudeFrom Longitude of start point in [deg decimal]
+     * @param float $latitudeTo    Latitude of target point in [deg decimal]
+     * @param float $longitudeTo   Longitude of target point in [deg decimal]
+     * @param float $earthRadius   Mean earth radius in [m]
+     *
+     * @return float Distance between points in [m] (same as earthRadius)
+     */
+    public static function vincentyGreatCircleDistance(
+        float $latitudeFrom,
+        float $longitudeFrom,
+        float $latitudeTo,
+        float $longitudeTo,
+        float $earthRadius = 6371000
+    ): float {
+        // convert from degrees to radians
+        $latFrom = deg2rad($latitudeFrom);
+        $lonFrom = deg2rad($longitudeFrom);
+        $latTo = deg2rad($latitudeTo);
+        $lonTo = deg2rad($longitudeTo);
+
+        $lonDelta = $lonTo - $lonFrom;
+        $a = (cos($latTo) * sin($lonDelta)) ** 2 +
+        (cos($latFrom) * sin($latTo) - sin($latFrom) * cos($latTo) * cos($lonDelta)) ** 2;
+        $b = sin($latFrom) * sin($latTo) + cos($latFrom) * cos($latTo) * cos($lonDelta);
+
+        $angle = atan2(sqrt($a), $b);
+
+        return $angle * $earthRadius;
+    }
+}